requirements.txt
aniso8601==9.0.1
attrs==23.2.0
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.2
flask-restplus==0.13.0
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
importlib_metadata==7.0.2
importlib_resources==6.1.3
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
marshmallow==3.21.1
nose==1.3.7
packaging==24.0
pkgutil_resolve_name==1.3.10
psycopg2==2.9.9
pytz==2024.1
referencing==0.33.0
rpds-py==0.18.0
Scaffold==0.1.3
six==1.16.0
SQLAlchemy==2.0.28
typing_extensions==4.10.0
Werkzeug==3.0.1
zipp==3.17.0
